refact: refactor main page and header
