+++ RSS-React/src/App.tsx
@@ -1,12 +1,12 @@
 import './App.css';
 import React from 'react';
-import Header from './components/header/header';
-import Footer from './components/footer/footer';
-import NotFoundPage from './components/pages/not-found-page/not-found-page';
-import AboutUsPage from './components/pages/about-us-page/about-us-page';
-import FormPage from './components/pages/form-page/form-page';
+import Header from './components/header/Header';
+import Footer from './components/footer/Footer';
+import NotFoundPage from './components/pages/not-found-page/NotFoundPage';
+import AboutUsPage from './components/pages/about-us-page/AboutUsPage';
+import FormPage from './components/pages/form-page/FormPage';
 import { Route, Routes } from 'react-router-dom';
-import MainPage from './components/pages/main-page/main-page';
+import MainPage from './components/pages/main-page/MainPage';
 
 class App extends React.Component {
   render(): React.ReactNode {
