+++ RSS-React/src/components/header/header.tsx
@@ -1,20 +1,11 @@
 import React from 'react';
-import { NavLink } from 'react-router-dom';
-import { withRouter } from '../../utils/withRouter';
+import { NavLink, useLocation } from 'react-router-dom';
 import styles from './header.module.scss';
 
-type LocationProps = {
-  location: {
-    pathname: string;
-  };
-};
+const Header: React.FC = () => {
+  const location = useLocation();
 
-class Header extends React.Component<LocationProps> {
-  constructor(props: LocationProps) {
-    super(props);
-  }
-
-  getCurrentPageName(currentURL: string): string {
+  const getCurrentPageName = (currentURL: string): string => {
     switch (currentURL) {
       case '/':
         return 'Main page';
@@ -31,29 +22,26 @@ class Header extends React.Component<LocationProps> {
         return 'Ooops...';
         break;
     }
-  }
+  };
 
-  render(): React.ReactNode {
-    return (
-      <header className={styles.wrapper}>
-        <nav className={styles.navigation}>
-          <ul className={styles.menu}>
-            <li className={styles.list}>
-              <NavLink to="./">Main page</NavLink>
-            </li>
-            <li className={styles.list}>
-              <NavLink to="./about-us">About us</NavLink>
-            </li>
-            <li className={styles.list}>
-              <NavLink to="./form">Form</NavLink>
-            </li>
-          </ul>
-        </nav>
-        <span className={styles.current_page}>
-          {this.getCurrentPageName(this.props.location.pathname)}
-        </span>
-      </header>
-    );
-  }
-}
-export default withRouter((props) => <Header {...props} />);
+  return (
+    <header className={styles.wrapper}>
+      <nav className={styles.navigation}>
+        <ul className={styles.menu}>
+          <li className={styles.list}>
+            <NavLink to="./">Main page</NavLink>
+          </li>
+          <li className={styles.list}>
+            <NavLink to="./about-us">About us</NavLink>
+          </li>
+          <li className={styles.list}>
+            <NavLink to="./form">Form</NavLink>
+          </li>
+        </ul>
+      </nav>
+      <span className={styles.current_page}>{getCurrentPageName(location.pathname)}</span>
+    </header>
+  );
+};
+
+export default Header;
